Conflicts:
	Dev/Dev2.Runtime.Services/ServiceModel/Esb/Brokers/ComPlugin/ComPluginRuntimeHandler.cs

+++ Dev/Dev2.Runtime.Services/ESB/Management/Services/MoveItemService.cs
@@ -66,8 +66,12 @@ namespace Dev2.Runtime.ESB.Management.Services
                 item = ServerExplorerRepo.MoveItem(itemToMove, newPath.ToString(), GlobalConstants.ServerWorkspaceID);
                 if(item.Status==ExecStatus.Success && itemToMove.ResourceType=="Folder")
                 {
-                    if (Directory.Exists(ServerExplorerRepository.DirectoryStructureFromPath(itemToMove.ResourcePath)))
-                          Directory.Delete(ServerExplorerRepository.DirectoryStructureFromPath(itemToMove.ResourcePath),true);
+                    if (itemToMove.Children != null && itemToMove.Children.Count > 0)
+                    {
+
+                        if (Directory.Exists(ServerExplorerRepository.DirectoryStructureFromPath(itemToMove.ResourcePath)))
+                            Directory.Delete(ServerExplorerRepository.DirectoryStructureFromPath(itemToMove.ResourcePath), true);
+                    }
                     //ServerExplorerRepo.Reload(Guid.Empty);
                 }
             }
